+++ src/kanban/components/todo/deleteTodo.js
@@ -1,17 +1,19 @@
-import { api } from '../../api/api.js';
-import { renderMemberTodo } from './renderMemberTodo.js';
+import { api } from '../../../api/api.js';
+import { renderTodo } from './renderTodo.js';
 
 export const deleteTodo = async (target, currentTeam) => {
   const teamId = currentTeam.id;
   const itemId = target.closest('li').id;
   const memberId = target.closest('.todoapp-container').id;
 
-  console.log(teamId, memberId, itemId);
+  /**
+   **  현재 Memeber의 Todo가 Delete가 안 됨
+   */
 
   try {
-    await api.deleteMemberTodo(teamId, memberId, itemId);
+    api.deleteMemberTodo(teamId, memberId, itemId);
     const member = await api.getMember(teamId, memberId);
-    renderMemberTodo(memberId, member.todoList);
+    renderTodo(memberId, member.todoList);
   } catch (err) {
     throw new Error(err);
   }
