+++ src/utils/constant.js
@@ -1,4 +1,4 @@
-export const BASE_URL = 'https://js-todo-list-9ca3a.df.r.appspot.com';
+export const BASE_URL = 'https://js-todo-list-9ca3a.df.r.appspot.com/';
 
 export const TEAM = Object.freeze({
   MIN_TEAM_NAME_LEN: 2,
